requirements.txt
absl-py==2.1.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
dill==0.4.0
executing==2.2.0
fastapi==0.115.12
ffmpeg==1.4
ffmpeg-python==0.2.0
flatbuffers==25.2.10
fonttools==4.56.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.68.1
h11==0.16.0
h5py==3.13.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
pillow_heif==0.20.0
platformdirs==4.3.6
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.50
protobuf==5.29.2
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytubefix==8.12.2
pytz==2025.2
pywin32==309
pyzmq==26.2.1
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.2
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yt-dlp==2025.2.19
